Check revision on commit (#41)
In the code editor there will be checks on refocus, whether the latest revision of the branch (or the repository if there are no branches) has changed. In this case, the user will get a warning and the commit button will stay disabled. If the user commits a file without noticing changes in the repository, the server will reject the change and a "concurrent modification error" will be shown.

This feature works with SCM-Manager 2.23.0 or later, but does not need this version. So the required SCM-Manager version has not been changed for this feature.

Co-authored-by: Konstantin Schaper <konstantin.schaper@cloudogu.com>