Improve vertical scrolling for smaller screens (#73)
On smaller screens, the horizontal scrollbar in the fullscreen view is only displayed when editing a file as soon as the user scrolls all the way down in the modal. Otherwise, the user can only navigate horizontally using a touchpad or keyboard. With this change, the vertical scrollbar should always be displayed.