Modal must not trigger tooltip (#78)
Change tooltip style to be no longer triggered by modal which lies directly over the tooltip. This happended because the modal is nested inside the tooltip wrapper and after refocusing the window the modal mouse interactions also triggered the tooltip active state. Since we already plan to create a new tooltip component I decided to take the workaround by changing the tooltip style. We also use the same style for the content action bar.